Refactor: ApplicationFormControl 페이지 CreateApplicationFormControl로 파일명 변경
