+++ src/pages/index.ts
@@ -1,3 +1,3 @@
 export { default as LoginPage } from './LoginPage/LoginPage.page';
-export { default as CreateApplicationForm } from './ApplicationFormControl/ApplicationFormControl.page';
+export { default as CreateApplicationForm } from './CreateApplicationForm/CreateApplicationForm.page';
 export { default as ApplicationFormDetail } from './ApplicationFormDetail/ApplicationFormDetail.page';
